splitting moving
