+++ packages/web/src/ScriptContext.tsx
@@ -1,16 +1,5 @@
-import React, { createContext, useState, useMemo, useEffect, use } from "react"
-import { ChatModels } from "../../core/src/chattypes"
-import {
-    Project,
-    ServerEnvResponse,
-    RunResultListResponse,
-} from "../../core/src/server/messages"
-import { fetchEnv, fetchScripts, fetchRuns, fetchModels } from "./api"
+import React, { createContext, use } from "react"
 import { useLocationHashValue } from "./useLocationHashValue"
-import { useUrlSearchParams } from "./useUrlSearchParam"
-import { FileWithPath } from "react-dropzone/."
-
-export type ImportedFile = FileWithPath & { selected?: boolean }
 
 export const ScriptContext = createContext<{
     scriptid: string | undefined
