+++ packages/web/src/types.ts
@@ -1,0 +1,3 @@
+import { FileWithPath } from "react-dropzone/."
+
+export type ImportedFile = FileWithPath & { selected?: boolean }
